rename files and add missing types
