+++ src/features/permissions/constants/permissions.keys.ts
@@ -1,0 +1,1 @@
+export const PERMISSIONS_KEY = "permissions";
