+++ src/features/permissions/permissions.ts
@@ -1,3 +1,4 @@
 export { default as Permission } from "./components/Permission/Permission";
-export * from "./logics/permissionLogic";
-export * from "./constants/permissionScopes";
+export { default as usePermissions } from "./hooks/permissions.hooks";
+export { setPermissions, clearPermissions } from "./logics/permissions.logics";
+export { PermissionEnum } from "./constants/permissions.scopes";
